requirements.txt
pyannote.audio
requests
openai-whisper
openai
pydub
minio
